race data communication

+++ Source/Core/Core/Falconnect/RacerMemoryBlock.h
@@ -12,7 +12,7 @@ public:
 
     std::vector<u8> GetSocketData() const;
 
-    std::vector<u8> GetDolphinPatchData(std::vector<u32> currentData);
+    std::vector<u32> GetDolphinPatchData(std::vector<u32> currentData);
 
     u32 state;
     u32 centerPosition[3];
